# perf: improved file opening speed in Windows