--- a/synfig-core/src/synfig/transform.cpp
+++ b/synfig-core/src/synfig/transform.cpp
@@ -32,6 +32,7 @@
 #	include <config.h>
 #endif
 
+#include "general.h"
 #include "transform.h"
 #include <algorithm>
 
@@ -64,13 +65,13 @@ TransformStack::perform(const synfig::Vector& x)const
 {
 	synfig::Vector ret(x);
 
-	if (getenv("SYNFIG_DEBUG_TRANSFORM_STACK")) printf("   PERFORM %s: %5.2f %5.2f", get_guid().get_string().substr(0,6).c_str(), ret[0], ret[1]);
+	DEBUG_LOG("SYNFIG_DEBUG_TRANSFORM_STACK", "   PERFORM %s: %5.2f %5.2f", get_guid().get_string().substr(0,6).c_str(), ret[0], ret[1]);
 	for(const_reverse_iterator iter(rbegin());iter!=rend();++iter)
 	{
 		ret=(*iter)->perform(ret);
-		if (getenv("SYNFIG_DEBUG_TRANSFORM_STACK")) printf(" (%14s) %5.2f %5.2f", (*iter)->get_string().c_str(), ret[0], ret[1]);
+		DEBUG_LOG("SYNFIG_DEBUG_TRANSFORM_STACK", " (%14s) %5.2f %5.2f", (*iter)->get_string().c_str(), ret[0], ret[1]);
 	}
-	if (getenv("SYNFIG_DEBUG_TRANSFORM_STACK")) printf("\n");
+	DEBUG_LOG("SYNFIG_DEBUG_TRANSFORM_STACK", "\n");
 
 	return ret;
 }
@@ -80,13 +81,13 @@ TransformStack::unperform(const synfig::Vector& x)const
 {
 	synfig::Vector ret(x);
 
-	if (getenv("SYNFIG_DEBUG_TRANSFORM_STACK")) printf(" UNPERFORM %s: %5.2f %5.2f", get_guid().get_string().substr(0,6).c_str(), ret[0], ret[1]);
+	DEBUG_LOG("SYNFIG_DEBUG_TRANSFORM_STACK", " UNPERFORM %s: %5.2f %5.2f", get_guid().get_string().substr(0,6).c_str(), ret[0], ret[1]);
 	for(const_iterator iter(begin());iter!=end();++iter)
 	{
 		ret=(*iter)->unperform(ret);
-		if (getenv("SYNFIG_DEBUG_TRANSFORM_STACK")) printf(" (%14s) %5.2f %5.2f", (*iter)->get_string().c_str(), ret[0], ret[1]);
+		DEBUG_LOG("SYNFIG_DEBUG_TRANSFORM_STACK", " (%14s) %5.2f %5.2f", (*iter)->get_string().c_str(), ret[0], ret[1]);
 	}
-	if (getenv("SYNFIG_DEBUG_TRANSFORM_STACK")) printf("\n");
+	DEBUG_LOG("SYNFIG_DEBUG_TRANSFORM_STACK", "\n");
 
 	return ret;
 }