@ice0 2022-09-12
> (PS: Are the results for MacOS inverted?)

Yes. Updated info, thanks.