@morevnaproject 2022-09-26
Backported to 1.4.x branch :heavy_check_mark:  - https://github.com/synfig/synfig/commit/8d7a690e059562cf64bbb1efdcc0bf434a0deb31